Add "setAngle" & "setFlip" (+getters)
setAngle can be called on an image to rotate it around the origin point (center). The arguent is a double in degrees, clockwise.

setFlip can be called on an image to flip it. It takes two bools, the first being whether or not it should be flipped on the X-axis. The second one is for the Y-axis. Both can be set to true at the same time, flipping the image in both the X- and Y-axis.